feat: reset scroll manually
