+++ src/pages/Trade/index.tsx
@@ -12,6 +12,7 @@ import iconSettingHover from "assets/icons/icon-setting-hover.svg";
 import useSettingsModal from "hooks/modals/useSettingsModal";
 import { MOBILE_SCREEN_CLASS } from "constants/layout";
 import Tooltip from "components/Tooltip";
+import ScrollToTop from "components/ScrollToTop";
 
 const Wrapper = styled.div`
   width: 100%;
@@ -29,6 +30,7 @@ function TradePage() {
   const settingsModal = useSettingsModal();
   return (
     <Wrapper>
+      <ScrollToTop />
       <Container>
         <Row justify="center" gutterWidth={0}>
           <Col width={512} xs={12} className="modal-parent">
